+++ packages/lib/src/components/internal/SecuredFields/lib/CSF/utils/cardType.ts
@@ -1,5 +1,5 @@
-import { CardObject } from '../types';
-import { hasOwnProperty } from '../../../../../utils/hasOwnProperty';
+import { CardObject } from '../../types';
+import { hasOwnProperty } from '../../../../../../utils/hasOwnProperty';
 
 let shortestPermittedCardLength;
