+++ packages/lib/src/components/internal/SecuredFields/lib/CSF/utils/userAgent.ts
